+++ src/progress.h
@@ -33,10 +33,10 @@ struct fwup_progress {
     // If we're showing progress when applying, this is the number of progress_units that
     // should be 100%. Each unit corresponds to about one byte of work. It's not intended
     // as an exact measure of bytes written.
-    int64_t total_units;
+    uint64_t total_units;
 
     // This counts up as we make progress.
-    int64_t current_units;
+    uint64_t current_units;
 
     // The most recent progress reported is cached to avoid unnecessary context switching/IO
     int last_reported_percent;
@@ -52,7 +52,7 @@ struct fwup_progress {
 };
 
 void progress_init(struct fwup_progress *progress, int progress_low, int progress_high);
-void progress_report(struct fwup_progress *progress, int progress_units);
+void progress_report(struct fwup_progress *progress, uint64_t progress_units);
 void progress_report_complete(struct fwup_progress *progress);
 
 #endif // PROGRESS_H
